+++ library/SolverUtils/DriverParareal.cpp
@@ -68,6 +68,14 @@ DriverParareal::~DriverParareal()
 /**
  *
  */
+void DriverParareal::v_InitObject(std::ostream &out)
+{
+    DriverParallelInTime::v_InitObject(out);
+}
+
+/**
+ *
+ */
 void DriverParareal::v_Execute(std::ostream &out)
 {
     // Set timer.