+++ library/SolverUtils/DriverStandard.cpp
@@ -73,6 +73,9 @@ void DriverStandard::v_InitObject(ostream &out)
     Driver::v_InitObject(out);
 }
 
+/**
+ *
+ */
 void DriverStandard::v_Execute(ostream &out)
 
 {
@@ -101,6 +104,7 @@ void DriverStandard::v_Execute(ostream &out)
     // Evaluate and output computation time and solution accuracy.
     // The specific format of the error output is essential for the
     // regression tests to work.
+
     // Evaluate L2 Error
     for (int i = 0; i < m_equ[0]->GetNvariables(); ++i)
     {